Add a first tree implementation
A first implementation of a tree program which only reads a directory as
argument and outputs the subdirectories in a tree like format with
special non-ASCII characters as the tree elements.

Use the `Write` trait so that either `stdout` can be passed or something
to save the output in, which makes unit testing easier.

The tree levels are saved as `Vector` elements which can be one of the
enum `TreeLevel`. This approach makes it easier to chose different ways
to render the tree later. The `Vector` also makes it easy to get the
current indentation depth, which is equal to the length.

Finally also add unittests, which temporarly create a directory tree to
check our implementation against. Also add another unittest to ensure
the proper rendering of the `TreeLevel` elements.

Signed-off-by: SteveCharleston <steven@schalhorn.org>