Add multithreading
Execute each visit of a directory in a new thread. This way we aren't
limited by the stack size when doing recursion into deeply neested
directory structures. Additionally we will be able to do expensive
computations in future in a less blocking fashion.

To keep the ordering we first spawn all threads and save them in a
`Vec`. Later join all threads in the order they have been spawned to get
the correct ordered output. Since parent threads will wait for child
threads in this way, we are essentially doing a depth-first search.

Signed-off-by: SteveCharleston <steven@schalhorn.org>