Add error handling in case of lacking permissions
If it is not possible to enter a directory due to missing permissions
for the current user or if the directory hasn insufficient permissions
we simply do not want to enter the directory and go on with the next
one.

Added code to handle this situation and an unittest to verify that.

Additionally removed the unnecessary `Writer` argument from all
functions which is unnecessary now.

Signed-off-by: SteveCharleston <steven@schalhorn.org>