Show message about missing permissions on dir line
To indicate missing permissions to enter a directory, we want to show an
error message. The function `print_paths` has been modified to not only
return a `String` but rather a `Result` containing the rendered subtree
or an error. The caller can then handle the error and insert an
appropriate message or simply use the returned subtree.

As a consequence the way how tree entries are rendered is changed to
insert the newline at the beginning of a line instead of at the end.
This way we can add text to a parent node, since only the child nodes
will have the newline character.

Signed-off-by: SteveCharleston <steven@schalhorn.org>