Change implementation to construct tree of structs
Instead of directly rendering the string representation of a tree a
struct is first used to represent the filesystem. The string
representation is then rendered afterwards from this data-structure.

This approach gives us more flexibility in extending the information we
want to process from the filesystem (e.g. file permissions, extended
attributes or self generated checksums), since we can now easily extend
the `TreeEntry` structure and use that information later when
appropriate. This also enables us to render the data in different ways
such as json or yaml in the future.

Signed-off-by: SteveCharleston <steven@schalhorn.org>