Add classify option
Implemented a first version of a classify option, as it is present in
`ls`. This version only appends `/` to directories, since only files and
directories are recognized as distinct filetypes in this commit.

This is a preparation for more options and modifications on the
rendering of the data. Therefore the command options parsed by `clap`
are moved into `lib.rs` to make it useable as general part of this
library. Also an `option` argument has been added to `tree` to pass
cli options in.

Also add unittests to verify the classify option.

Signed-off-by: SteveCharleston <steven@schalhorn.org>