Add colorization of tree entries
Colorize the output according to `LS_COLORS` environment variable in the
same way that `ls` colors directory entries. Use the `lscolors` crate to
calculate the coloring and ANSI sequences.

To make this possible change the `name` entry in a `TreeEntry` to a
`PathBuf` from a `String`. The name can easily be extracted from that,
just as before, but we also get the full path to the file as well as an
easy way to read the metadata.

Also add an option to turn colorization off as well as tests to verify
that this flag is honored and text is colored.

Signed-off-by: SteveCharleston <steven@schalhorn.org>