Count files and directories
Add a report at the end of the tree output to show how many files and
directories have been seen. Add a new option `noreport` to disable this
report.

Add an unittest which creates several files and directories and verify
that the report is correct after each new file or directoy.

Signed-off-by: SteveCharleston <steven@schalhorn.org>