Limit directory depth if requested
Add the `--limit` flag to set the recursion limit into the directory
tree. Add an unittest which checks that a directory is not listed if the
level is not high enough and that the directory is listed if it is.

Level `0` means displaying the whole tree which contrasts with the
original unix tree that just aborts in that case.

Signed-off-by: SteveCharleston <steven@schalhorn.org>