Add means to calculate longest field of TreeEntry
Add a function to calculate the longest node in the recursive list of
TreeEntry structs. To do that we supply a function which is able to
retrieve the string representation of the field we are interested in.

This will be used to help with the formatting of fields that will be
introduced later.

Signed-off-by: SteveCharleston <steven@schalhorn.org>