Add option to list user or group metadata
Make it possible to display the user and/or group information for the
`TreeEntry` nodes. Use the newly introduces function `longest_fieldentry`
to calculate the node with the longest string representation of group or
user. Use this information to uniformly format all `TreeEntry` nodes in
the same way, so that the filenames appear in the same column, even
though user/group data may differ in length.

Introduce a new struct `TreeLevelMeta`, which for now holds the
user/group information but will be extended in the future.

Signed-off-by: SteveCharleston <steven@schalhorn.org>