Improve error displaying
Show the error message, that a directory could not be accessed, in a new
separeate line. Also colorize that error message in red.

Fix an existing bug where the program crashed if there are insufficient
permissions to access the root dir of the tree invocation target.

As a consequence change the `recurse_paths` function to return a
`TreeChild` instance instead of a `Result` with either a valid
`TreeEntry` list or an error. This way we do not have to distinguish
those two cases later, since `TreeChild::Error` is a valid result that
could be stored directly.

Signed-off-by: SteveCharleston <steven@schalhorn.org>