Refactor TreeLevelMeta::from to take Option struct
Since the metadata attributes we want to gather from a file will only
grow, refactor `TreeLevelMeta::from` now to take an `Option` struct and
get the metadata to store of a file directly from there. This way the
function attributes do not need to be extended every time we want to
store more information of a file.

Signed-off-by: SteveCharleston <steven@schalhorn.org>