Refactor fixture create_directory_tree with loops
Do not call the directory and file creation functions for all files and
directories in `create_directory_tree`. Instead create a list of those
and loop over the lists to create them. This makes extending the
directory tree easier and the function becomes clearer. It is also
easier to automatically to some more transformations for those files and
directories, that might benefit tests that would be implemented later.

Signed-off-by: SteveCharleston <steven@schalhorn.org>