Add display of file sizes
Display the size of a file in bytes using the `sizes` option.
Additionally add an option `humansize` to display the size better
readable in KiB, MiB and so on. Use the crate `humansize` for that,
since this crate ships with a formatter option, that can easily be
changed to use SI units later.

Signed-off-by: SteveCharleston <steven@schalhorn.org>