Add handling of symlinks
Symlinks are now displayed and followed correctly. Broken links are
recogized and colored as a missing file, working links are colored with
the correct file type of the target file. Displaying of symlinks is done
with an arrow sign on the same line as the symlink itself.

Symlinks to files and directories are also counted correctly in the
final report, dangling symlinks are also correctly counted as files.

Before that change symlinks where followed like any other directory,
which could have resulted in infinite recursion, in case of loops in a
symlink structure.

Signed-off-by: SteveCharleston <steven@schalhorn.org>