Add a filter option with glob pattern
Provide the option `--parser` which specifies a filter to show only
files that match the given pattern. This supports normal globs with the
help of the `GlobSet` library. Multiple globs can be separated by a pipe
character '|'. This conforms with the original tree implementation where
globs could be separated by a pipe as well.

Signed-off-by: SteveCharleston <steven@schalhorn.org>