Fix indentation on permission errors
When additional metadata is displayed, like chmods or filesize, the
error message has to be indented further to account for the additional
horizontal space taken by the parent entry.

To achieve that the length of the string with the metadata is taken,
before the actual tree lines and the file name is rendered.

Signed-off-by: SteveCharleston <steven@schalhorn.org>