Add option to not cross filesystem boundaries
Use the option `--xdev` to check if we are leaving the filesystem, where
the root path is. Do that with the FS attribute `st_dev`, that is returned by
a libc `stat` call.

Since it is not really possible to test such a feature, without complex
mocking, we simply check that the tree is rendered correctly when this
option is enabled.

Signed-off-by: SteveCharleston <steven@schalhorn.org>