Add option to follow symlinks
To reliably follow symlinks, it is necessary to also detect if we follow
a loop in the filesystem. Otherwise the program would never stop and
progressively consume all CPU and RAM.

To achieve that a set of already seen inodes on a device is carried in
each recursive call. This set is cloned and the new inode is added
before each recurisve call, therefore we don't have to worry about
shared resources and locking. It is compared if the set of seen inodes
changed from the last call, to decide if we really face a loop. Most
loops should be detected this way.

The set of inodes is wrapped in an `Option` which is simply `None` if we
have not activated following of symlinks. Besides the command line flag,
this `Option` is also an indicator if we want to follow symlinks.

Also add unittests that have looping, to test if this feature works.

Signed-off-by: SteveCharleston <steven@schalhorn.org>