Harmonize indendations
The last branch of a tree or subtree was always indendet too less. While
this looked a little bit nicer, it made it way harder to distinguish the
levels of all files as a whole. Therefore increase the indendation of
the last (sub)-tree to make all the entry indendations the same.

Signed-off-by: SteveCharleston <steven@schalhorn.org>