Handle non-readable directory contents
It is sometimes possible to get an iterator of the contents of a
directory but then to not read out the entries. That might happen e.g.
in case of a network filesystem that goes offline or if reading the
contents of a zombie process inside the `/proc` directory, among other
things.

Due to this the contents of a directory are first collected completly
and only afterwards filtered. A common trick is used, where `collect`
can store a Vector inside a `Result`, which could also store `Err` in
case something went wrong.

Signed-off-by: SteveCharleston <steven@schalhorn.org>