Output to a Writer instead of using a String
Using a string to store the tree representation could result in high
memory usage. Especially in the case of using `--cat` to output files
content.

Store the tree output in a `Writer` instead, which makes it possible to
also supply `stdout` and therefore directly print everything without
storing anything in memory.

Signed-off-by: SteveCharleston <steven@schalhorn.org>