Add filetype detection
Detect the mime type of a file using the `tree_magic` crate and storage
that information in the `TreeLevelMeta` struct if requested.

Add a special handling of dangling symlinkgs, since those are not
recognized otherwise. The logic of `tree_magic` works with looking at
the first bytes of a file, which would not work without a file to read
from.

Also fix a long existing bug where metadata of symlinkgs would not be
read and instead the metadata of the target file would be requested.
This would not work at all on dangling symlinks which caused display
errors.

Signed-off-by: SteveCharleston <steven@schalhorn.org>