Add filetype filter option
Make it possible to filter tree entries by their filetype. To achieve
this it is necessary to possible scan a file twice for its filetype.
Once when iterating over a directory and filtering the unwanted entries
and then later on if displaying of a filetype is requested. However
there is no sensible way to filter tree entries after the directory
scan, since otherwise the tree indents would have to be recursively
adjusted for parent nodes.

Signed-off-by: SteveCharleston <steven@schalhorn.org>