Move filetype filter into TreeEntry generation
Because the `TreeEntry` levels are recalculated as a separate step, it
is now possible to do complex filter operations while construction the
initial `TreeEntry` list.

Due to this the filtering of the filetypes has been moved into this
stage. A double calculation of the filetype can be avoided this way and
the code became less complex.

It was necessary to exchange the `map` with a `filter_map` to exclude
the unwanted directory entries from the generated list.

This enables us to also do more complex filtering in the future, with
other critera.

Signed-off-by: SteveCharleston <steven@schalhorn.org>