Add option to prune empty dirs
The new option `prune` can be used to filter out empty directories from
the output. This is especially usefull in combination with the filetype
or pattern filter options, which might leave behind a huge amount of
empty directories after the filter was applied. With the `prune` option
only the subtrees with matched files are now shown, resulting in a much
better overview of the relevant directory tree parts.

Signed-off-by: SteveCharleston <steven@schalhorn.org>