Add option to execute a command on all files
The new option `exec` has been added, which executes the given command
on each and every file in the tree. The command output is added under
the file node in the tree, in the same way it is done in the `cat`
command. The string `{}` is replaced in a given command string, which
has been inspired by the unix `find` command. However unlike `find` the
command to execute has to be passed as one single argument.

An example call to the programm could be `--exec 'ls -l {}'`.

Signed-off-by: SteveCharleston <steven@schalhorn.org>