Run exec commands in multiple threads
Move the execution  of the external program into the instantiation of
`TreeLevelMeta`, so that the program is executed directly when every
file is encountered. Since the walk through the filesystem is multi
threaded, the execution is parallized as well.

Additionally introduce a struct `ExecCommand` which holds a program name
and the arguments to be executed and also introduce a function
`ExecCommand::from` to parse a string into that struct.

Signed-off-by: SteveCharleston <steven@schalhorn.org>